+++ tests/Integration/Authentication/TokenAuthenticatorTest.php
@@ -6,15 +6,18 @@ use Mockery;
 use GuzzleHttp\Client;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
+use GuzzleHttp\Exception\ClientException;
 use Illuminate\Contracts\Cache\Repository;
 use EthicalJobs\SDK\Authentication\TokenAuthenticator;
 use EthicalJobs\SDK\Testing\ResponseFactory;
+
+
+use Tests\Helpers\MockResponseStack;
 
 class TokenAuthenticatorTest extends \Tests\TestCase
 {
     /**
      * @test
-     * @group Unit
      */
     public function it_can_set_its_credentials_and_get_its_token_from_cache()
     {
@@ -32,7 +35,7 @@ class TokenAuthenticatorTest extends \Tests\TestCase
             ->once()
             ->withArgs([
                 'POST',
-                'http://api.ethicaljobs.com.au/oauth/token',
+                'https://api.ethicaljobs.com.au/oauth/token',
                 [
                     'json' => [
                         'grant_type'    => 'password',
@@ -58,15 +61,16 @@ class TokenAuthenticatorTest extends \Tests\TestCase
                     return true;
                 }),
             ])
+            ->andReturn('mock-token-aks-sjs-38w')
             ->getMock();  
 
-        (new TokenAuthenticator($client, $cache, $credentials))
-            ->getToken();
+        $token = (new TokenAuthenticator($client, $cache, $credentials))->getToken();
+
+        $this->assertEquals('mock-token-aks-sjs-38w', $token);
     }    
 
     /**
      * @test
-     * @group Unit
      */
     public function it_sets_an_authorization_bearer_header()
     {
@@ -91,4 +95,79 @@ class TokenAuthenticatorTest extends \Tests\TestCase
 
         $this->assertEquals($expected, $authenticated->getHeaders());
     }        
+
+    /**
+     * @test
+     */
+    public function it_throws_correct_exception_on_401()
+    {
+        $this->expectException(ClientException::class);
+
+        $response = ResponseFactory::response(401, ResponseFactory::authentication());
+
+        $request = new Request('GET', 'https://github.com/stars');
+
+        MockResponseStack::mock([
+            new ClientException('Unauthorized', $request, $response),        
+        ]);
+
+        $middleware = resolve(TokenAuthenticator::class);
+
+        try {
+            $middleware->authenticate($request);
+        } catch (\Exception $exception) {
+            $this->assertEquals('Unauthorized SDK Request', $exception->getMessage());
+            throw $exception; // Dont swallow exceptions in tests - leads to false positives.
+        }
+    }      
+
+    /**
+     * @test
+     */
+    public function it_throws_correct_exception_on_403()
+    {
+        $this->expectException(ClientException::class);
+
+        $response = ResponseFactory::response(403, ResponseFactory::authentication());
+
+        $request = new Request('GET', 'https://github.com/stars');
+
+        MockResponseStack::mock([
+            new ClientException('Unauthorized', $request, $response),        
+        ]);
+
+        $middleware = resolve(TokenAuthenticator::class);
+
+        try {
+            $middleware->authenticate($request);
+        } catch (\Exception $exception) {
+            $this->assertEquals('Unauthorized SDK Request', $exception->getMessage());
+            throw $exception; // Dont swallow exceptions in tests - leads to false positives.
+        }
+    }    
+    
+    /**
+     * @test
+     */
+    public function it_does_not_modify_non_auth_exceptions()
+    {
+        $this->expectException(ClientException::class);
+
+        $response = ResponseFactory::response(404, ResponseFactory::authentication());
+
+        $request = new Request('GET', 'https://github.com/stars');
+
+        MockResponseStack::mock([
+            new ClientException('Unauthorized', $request, $response),        
+        ]);
+
+        $middleware = resolve(TokenAuthenticator::class);
+
+        try {
+            $middleware->authenticate($request);
+        } catch (\Exception $exception) {
+            $this->assertEquals('Unauthorized', $exception->getMessage());
+            throw $exception; // Dont swallow exceptions in tests - leads to false positives.
+        }
+    }      
 }